# server/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Food Delivery API",
    "description": "API for food delivery system including user, vendor, delivery agent, and analytics functionality",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5050/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
   "/menu/getVendors": {
  "post": {
    "tags": ["Menu"],
    "summary": "Get vendors with filtering options",
    "description": "Retrieves vendors along with their menu items using filters. Uses POST to handle complex filter criteria.",
    "operationId": "getVendors",
    "security": [{"bearerAuth": []}],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/VendorFilter"
          },
          "example": {
            "category": ["Veg"],
            "cuisine": ["Gujarati", "South Indian"],
            "priceSort": 1
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Vendors retrieved successfully",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/VendorsResponse"
            },
            "example": {
              "success": true,
              "message": "All vendors with their menus retrieved successfully",
              "vendors": [
                {
                  "_id": "68132e745ba22849c6499ded",
                  "firstname": "kevin",
                  "lastname": "patel",
                  "number": 9876512210,
                  "email": "kevin.doe@example.com",
                  "totalTables": 20,
                  "isActive": true,
                  "createdAt": "2025-05-01T08:19:00.210Z",
                  "updatedAt": "2025-05-07T10:59:01.298Z",
                  "fullname": "kevin patel",
                  "menu": [
                    {
                      "_id": "68133114ecf96a7ae0a0cc07",
                      "vendorId": "68132e745ba22849c6499ded",
                      "type": "Lunch",
                      "name": "rice",
                      "price": 1500,
                      "image": "/item/1746088213384.png",
                      "category": "Veg",
                      "cuisine": "Gujarati",
                      "createdAt": "2025-05-01T08:30:12.597Z",
                      "updatedAt": "2025-05-01T08:30:12.597Z"
                    },
                    {
                      "_id": "6813311cecf96a7ae0a0cc09",
                      "vendorId": "68132e745ba22849c6499ded",
                      "type": "Lunch",
                      "name": "thali",
                      "price": 1500,
                      "image": "/item/1746088228897.png",
                      "category": "Veg",
                      "cuisine": "Gujarati",
                      "createdAt": "2025-05-01T08:30:20.095Z",
                      "updatedAt": "2025-05-01T08:30:20.095Z"
                    }
                  ]
                }
              ]
            }
          }
        }
      },
      "400": {
        "description": "Invalid request parameters"
      },
      "401": {
        "description": "Unauthorized - Invalid or missing token"
      },
      "500": {
        "description": "Internal server error"
      }
    }
  }
},
...
"/order/GenerateBillByTable": {
    "post": {
      "tags": ["Order"],
      "summary": "Generate bill for a table and update payment status",
      "description": "Completes the order for a specific table and generates the final bill with payment details.",
      "operationId": "generateBillByTable",
      "security": [
        {
          "bearerAuth": []
        }
      ],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenerateBillRequest"
            },
            "example": {
              "tableNo": 6,
              "paymentStatus": "Paid"
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Bill generated successfully",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillResponse"
              },
              "example": {
                "success": true,
                "message": "Order for Table 6 completed successfully.",
                "bill": {
                  "tableNo": 6,
                  "orderId": "681afda633386dd2057cc1ca",
                  "items": [
                    {
                      "menuItemId": {
                        "_id": "68133114ecf96a7ae0a0cc07",
                        "vendorId": "68132e745ba22849c6499ded",
                        "type": "Lunch",
                        "name": "Rice",
                        "price": 1500,
                        "image": "/item/1746088213384.png",
                        "category": "Veg",
                        "cuisine": "Gujarati",
                        "createdAt": "2025-05-01T08:30:12.597Z",
                        "updatedAt": "2025-05-01T08:30:12.597Z"
                      },
                      "quantity": 2,
                      "price": 3000,
                      "status": "Served"
                    },
                    {
                      "menuItemId": {
                        "_id": "6813311cecf96a7ae0a0cc09",
                        "vendorId": "68132e745ba22849c6499ded",
                        "type": "Lunch",
                        "name": "Thali",
                        "price": 1500,
                        "image": "/item/1746088228897.png",
                        "category": "Veg",
                        "cuisine": "Gujarati",
                        "createdAt": "2025-05-01T08:30:20.095Z",
                        "updatedAt": "2025-05-01T08:30:20.095Z"
                      },
                      "quantity": 1,
                      "price": 1500,
                      "status": "Served"
                    }
                  ],
                  "totalAmount": 1507500,
                  "paymentStatus": "Paid",
                  "orderStatus": "Completed",
                  "createdAt": "2025-05-07T06:28:54.951Z",
                  "completedAt": "2025-05-07T06:36:47.065Z"
                }
              }
            }
          }
        },
        "400": {
          "description": "Invalid request - missing table number or invalid payment status"
        },
        "401": {
          "description": "Unauthorized - invalid or missing authentication token"
        },
        "404": {
          "description": "Table not found or no active order for the table"
        },
        "500": {
          "description": "Internal server error"
        }
      }
    }
  },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}